#12 feat: slide 컴포넌트 분리
