+++ src/App.tsx
@@ -3,13 +3,16 @@ import Router from './router/Router';
 import { ThemeProvider } from 'styled-components';
 import DefaultTheme from './styles/Theme';
 import GlobalStyle from './styles/GlobalStyle';
+import { RecoilRoot } from 'recoil';
 
 function App() {
   return (
-    <ThemeProvider theme={DefaultTheme}>
-      <GlobalStyle />
-      <Router />
-    </ThemeProvider>
+    <RecoilRoot>
+      <ThemeProvider theme={DefaultTheme}>
+        <GlobalStyle />
+        <Router />
+      </ThemeProvider>
+    </RecoilRoot>
   );
 }
