+++ src/recoil/slide.ts
@@ -1,0 +1,4 @@
+import { atom } from 'recoil';
+
+export const countState = atom({ key: 'count', default: 0 });
+export const hoverState = atom({ key: 'hover', default: true });
